refactor(file-name) Change discord-types.ts to discordTypes for consistency
